requirements.txt
pyalpm
python-pyfuse3 
